Added creating new subtasks, added confirmation window when removing tasks, added resetting forms when creating tasks

+++ react-front/src/components/task.tsx
@@ -1,4 +1,4 @@
-import { ComponentProps } from "react"
+import { ComponentProps, FormEvent, useRef, useState } from "react"
 import Subtask from "./subtask"
 import api from "../api/axios"
 
@@ -9,26 +9,49 @@ interface TaskProps extends ComponentProps<"article"> {
 }
 
 const Task = (props: TaskProps) => {
-
-    const task = props.task
-    // console.log(task.title);
+    const [task, setTask] = useState(props.task)
 
     const removeClicked = async () => {
         // TODO ask to confirm
+        const confirmed = window.confirm(`Remove task \'${task.title}\'?`)
+        if (!confirmed) return
+
         const resp = await api.delete(`/api/tasks/${props.task.id}`)
         props.onTasksUpdated(resp.data)
+    }
+
+    const addNewSubtask = async (e: FormEvent) => {
+        e.preventDefault()
+
+        const target = e.target as typeof e.target & {
+            title: { value: string },
+        }
+
+        const resp = await api.post('/api/tasks/subtask/add', {
+            title: target.title.value,
+            taskID: props.task.id
+        })
+        setTask(resp.data)
+
+        target.title.value = ''
     }
     
     return <article style={{border: "1px solid black", padding: 10}}>
         <h3>{task.title}</h3>
         <p style={{paddingLeft: 8}}>{task.description}</p>
-        <ul>
-            {task.subtasks.map(st => (
-                <li key={st.id.toString()}>
-                    <Subtask ownerTaskID={task.id} subtask={st} />
-                </li>
-            ))}
-        </ul>
+        <div>
+            <ul>
+                {task.subtasks.map(st => (
+                    <li key={st.id.toString()}>
+                        <Subtask ownerTaskID={task.id} subtask={st} />
+                    </li>
+                ))}
+                <form onSubmit={addNewSubtask}>
+                    <input type="text" name="title"/>
+                    <input type="submit" value="Add subtask"/>
+                </form>
+            </ul>
+        </div>
         <input type="button" value="Remove" onClick={removeClicked}/>
     </article>
 }
